etl-service/etl-project/requirements.txt
python-dotenv
pymongo
psycopg2-binary
